My suggestions to improve the project description
Question 5 about the grades requires to solve the data set problem and I don't think this is doable with what we cover in class! Asking the questions makes them feel they have to solve that. We cover defining schemas in project 3 so this is asking them about things that come later.